+++ lib/concurrent/executor/java_single_thread_executor.rb
@@ -1,6 +1,7 @@
 if Concurrent.on_jruby?
 
-  require 'concurrent/executor/executor_service'
+  require 'concurrent/executor/java_executor_service'
+  require 'concurrent/executor/serial_executor_service'
 
   module Concurrent
